twitoff/requirements.txt
astroid==2.2.5
basilica==0.2.6
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.4
oauthlib==3.0.1
Pillow==6.0.0
pylint==2.3.1
PySocks==1.7.0
python-decouple==3.1
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.5
tweepy==3.7.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
psycopg2==2.8.3
